Moved OAuth2RestClient and SCIM2RestClient to a separate package
